Changed one more int to long.

+++ src/edu/ysu/itrace/filters/Fixation.java
@@ -4,7 +4,7 @@ public class Fixation {
 	private RawGaze rawGaze;
 	private long duration;
 	
-	public Fixation(RawGaze rawGaze, int duration) {
+	public Fixation(RawGaze rawGaze, long duration) {
 		this.rawGaze = rawGaze;
 		this.duration = duration;
 	}
